optimization: improve tolerance breaking condition in gradient_descent
It does not make a big sense to check
	step * simulation.Dj_norm < tolerance
since this condition more or less repeats the loop breaking condition.

On the other hand, simulation.Dj_norm does not necessarily has to
converge to zero if a boundary constraint becomes active. Therefore it
is better to check
	simulation.PDj_norm < tolerance
(with no step coefficient).