Re-implement problem coefficients initialization
Previously, a Spline object must have known about the problem in order
to be callable with a dolfin Function as an argument.

The dolfin related code in splines module has been redisigned and moved
to a separate uflspline module.

The vhc, kappa_rad, and kappa_ax coefficients of the equation are now
initialized in problem as UFLSpline objects.